refactoring form index.jsx into two components
